SQL Project Code
In this code, I walk through various real life tasks from a movie dataset containing information including customers, movie rentals, and staff data. This is the beginning steps of my journey with SQL so this code does not include things joining with other tables . This code was done in MySQL Workbench, but should be very transferrable to other SQL versions.